# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nexo",
    "description": "API do Nexo. Este documento é o contrato: o cliente TypeScript do front é gerado a partir dele, nunca escrito à mão.",
    "version": "v1"
  },
  "paths": {
...
    "/pessoas": {
      "get": {
        "tags": [
          "Pessoas"
        ],
        "summary": "Lista as pessoas do tenant",
        "operationId": "ListarPessoas",
        "parameters": [
          {
            "name": "busca",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "papel",
            "in": "query",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Papel"
                }
              ]
            }
          },
          {
            "name": "incluirInativos",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "pagina",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "tamanho",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          },
          {
            "name": "ordenarPor",
            "in": "query",
            "schema": {
              "enum": [
                "Nome",
                "Codigo"
              ],
              "type": "string",
              "allOf": [
                {
                  "$ref": "#/components/schemas/OrdemDePessoas"
                }
              ],
              "default": 1
            }
          },
          {
            "name": "direcao",
            "in": "query",
            "schema": {
              "enum": [
                "Crescente",
                "Decrescente"
              ],
              "type": "string",
              "allOf": [
                {
                  "$ref": "#/components/schemas/Direcao"
                }
              ],
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaDePessoas"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pessoas"
        ],
        "summary": "Cria um cadastro",
        "operationId": "CriarPessoa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DadosDePessoa"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PessoaDetalhada"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaComProblemas"
                }
              }
            }
          }
        }
      }
    },
    "/pessoas/{id}": {
      "get": {
        "tags": [
          "Pessoas"
        ],
        "summary": "Abre um cadastro",
        "operationId": "ObterPessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PessoaDetalhada"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Pessoas"
        ],
        "summary": "Altera um cadastro",
        "operationId": "AlterarPessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/DadosDePessoa"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PessoaDetalhada"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaComProblemas"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Pessoas"
        ],
        "summary": "Inativa um cadastro",
        "description": "Não apaga: marca como inativo. Quem já apareceu em contrato ou cobrança precisa continuar existindo no histórico. Recusa enquanto houver contrato não encerrado ou cobrança em aberto.",
        "operationId": "InativarPessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaComProblemas"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/pessoas/{id}/reativar": {
      "post": {
        "tags": [
          "Pessoas"
        ],
        "summary": "Desfaz a inativação",
        "description": "Existe como endereço próprio, e não como um PUT com ativo=true, porque o PUT reescreve o cadastro inteiro: a tela precisaria mandar de volta campos que ela não tem em mãos, e apagaria o que não conhece.",
        "operationId": "ReativarPessoa",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DadosDePessoa": {
        "required": [
          "ativo",
          "regimeTributario",
          "tipo"
        ],
        "type": "object",
        "properties": {
          "tipo": {
            "enum": [
              "Fisica",
              "Juridica"
            ],
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/TipoPessoa"
              }
            ]
          },
          "regimeTributario": {
            "enum": [
              "Mei",
              "SimplesNacional",
              "LucroPresumido",
              "LucroReal",
              "TerceiroSetor",
              "PessoaFisica"
            ],
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/RegimeTributario"
              }
            ]
          },
          "responsavel": {
            "type": "string",
            "nullable": true
          },
          "papeis": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Papel"
            },
            "nullable": true
          },
          "nome": {
            "type": "string",
            "nullable": true
          },
          "nomeFantasia": {
            "type": "string",
            "nullable": true
          },
          "documento": {
            "type": "string",
            "nullable": true
          },
          "inscricaoEstadual": {
            "type": "string",
            "nullable": true
          },
          "inscricaoMunicipal": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "telefone": {
            "type": "string",
            "nullable": true
          },
          "celular": {
            "type": "string",
            "nullable": true
          },
          "endereco": {
            "allOf": [
              {
                "$ref": "#/components/schemas/DadosDeEndereco"
              }
            ],
            "nullable": true
          },
          "observacoes": {
            "type": "string",
            "nullable": true
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "Direcao": {
        "enum": [
          "Crescente",
          "Decrescente"
        ],
        "type": "string"
      },
...
      "OrdemDePessoas": {
        "enum": [
          "Nome",
          "Codigo"
        ],
        "type": "string"
      },
...
      "PaginaDePessoas": {
        "required": [
          "itens",
          "pagina",
          "tamanho",
          "total"
        ],
        "type": "object",
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PessoaNaLista"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "pagina": {
            "type": "integer",
            "format": "int32"
          },
          "tamanho": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "Papel": {
        "enum": [
          "Cliente",
          "Fornecedor",
          "Vendedor",
          "Colaborador"
        ],
        "type": "string"
      },
...
      "Problema": {
        "required": [
          "campo",
          "descricao",
          "sugestao",
          "titulo"
        ],
        "type": "object",
        "properties": {
          "campo": {
            "type": "string"
          },
          "titulo": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "sugestao": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "RegimeTributario": {
        "enum": [
          "Mei",
          "SimplesNacional",
          "LucroPresumido",
          "LucroReal",
          "TerceiroSetor",
          "PessoaFisica"
        ],
        "type": "string"
      },
...
      "RespostaComProblemas": {
        "required": [
          "problemas"
        ],
        "type": "object",
        "properties": {
          "problemas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Problema"
            }
          }
        },
        "additionalProperties": false
      },
...
      "TipoPessoa": {
        "enum": [
          "Fisica",
          "Juridica"
        ],
        "type": "string"
      }
    }
  }
}